Delete test1_1.csv
This was for my test of the reduced and radical species i.e 2nd order model, when I was trying to develop a way to get flipped/stacked in. I sent you this to troubleshoot one day, so It should not be uploaded with the paper or publicly available GitHub as its not being used nor is correct to use with 1st order rate constants.